Updated Downloader and DB with Last DL attempt date.
New Query added to db.js and existing queries modified so that the last_dl_attempt is taken into consideration when fetching app data to give to the downloader
retriver db functions initialise last_dl_attempt to 'epoch' so that the downloader will try to download new apps before trying previously failed ones.
downloadAPK.js renamed to downloader.js
downloader.js now queies for apps that haven't been downloaded and haven't been attempted to download for the most amount of time.
