+++ db/init_db.sql
@@ -4,14 +4,15 @@ create table apps(
 );
 
 create table app_versions(
-  id           serial         primary key not null,
-  app            text references apps(id) not null,
-  store          text                     not null,
-  region         text                     not null,
-  version        text                     not null,
-  screen_flags    int                             ,
-  downloaded     bool                     not null,
-  analyzed       bool                     not null
+  id              serial         primary key not null,
+  app               text references apps(id) not null,
+  store             text                     not null,
+  region            text                     not null,
+  version           text                     not null,
+  screen_flags       int                             ,
+  downloaded        bool                     not null,
+  analyzed          bool                     not null,
+  last_dl_attempt   timestamp                not null
 );
 
 create table developers(
